Move user utils to store
