+++ src/store/currentUser.ts
@@ -1,5 +1,6 @@
+import { RoleType } from '@beabee/beabee-common';
+import { computed, ComputedRef, ref } from 'vue';
 import axios from '../lib/axios';
-import { ref } from 'vue';
 import { fetchContact } from '../utils/api/contact';
 import { GetContactData } from '../utils/api/api.interface';
 
@@ -23,3 +24,13 @@ axios.interceptors.response.use(
     return Promise.reject(error);
   }
 );
+
+export const currentUserCan = (role: RoleType): ComputedRef<boolean> => {
+  return computed(() => {
+    return (
+      currentUser.value != null && currentUser.value.activeRoles.includes(role)
+    );
+  });
+};
+
+export const canAdmin = currentUserCan('admin');
