+++ src/store/index.ts
@@ -1,10 +1,12 @@
-import { currentUser, initCurrentUser, updateCurrentUser } from './currentUser';
-import {
-  generalContent,
-  initGeneralContent,
-  backgroundStyle,
-} from './generalContent';
+import { initCurrentUser } from './currentUser';
+import { initGeneralContent } from './generalContent';
 
 export const initStore = Promise.all([initCurrentUser, initGeneralContent]);
 
-export { currentUser, updateCurrentUser, generalContent, backgroundStyle };
+export { generalContent, backgroundStyle } from './generalContent';
+export {
+  currentUser,
+  updateCurrentUser,
+  currentUserCan,
+  canAdmin,
+} from './currentUser';
